store airwatch admin user id, remove it during user unprovisioning

+++ app/workers/custom_unprovision_worker.rb
@@ -4,6 +4,7 @@ class CustomUnprovisionWorker
   def perform(invitation_id)
     @invitation = Invitation.with_deleted.find_by_id(invitation_id)
     airwatch_id = @invitation.airwatch_user_id
+    airwatch_admin_id = @invitation.airwatch_admin_user_id
 
     puts "Will start custom un-provisioning for #{@invitation.recipient_username}:#{@invitation.recipient_email}"
     
@@ -41,40 +42,60 @@ class CustomUnprovisionWorker
     api_password = "$[5Jd#V7F."
     aw_tenant_code = "1LGKA4AAAAG5A5DACEAA"
 
-    #Call AirWatch API
-    begin
-      request_url = "https://apple.awmdm.com/API/v1/system/users/#{airwatch_id}/deactivate"
-      response = RestClient::Request.execute( :method => :post, 
-  									    	:url => request_url, 
-  									    	:user => "#{api_user}", 
-  									    	:password => "#{api_password}", 
-  									    	:headers => { :host => "apple.awmdm.com", 
-  									    		:authorization => "Basic bW9oYW46bW9oYW4=", 
-  									    		'aw-tenant-code' => "#{aw_tenant_code}"
-  									    	}
-    	)
+    if !airwatch_id.nil?
+      #Call AirWatch API
+      begin
+        request_url = "https://apple.awmdm.com/API/v1/system/users/#{airwatch_id}/deactivate"
+        response = RestClient::Request.execute( :method => :post, 
+    									    	:url => request_url, 
+    									    	:user => "#{api_user}", 
+    									    	:password => "#{api_password}", 
+    									    	:headers => { :host => "apple.awmdm.com", 
+    									    		:authorization => "Basic bW9oYW46bW9oYW4=", 
+    									    		'aw-tenant-code' => "#{aw_tenant_code}"
+    									    	}
+      	)
 
-      puts "AirWatch Account Deactivation for User #{@invitation.recipient_username}, AirWatch ID #{airwatch_id}. Response Code: #{response.code}"
-    rescue => e
-      puts e
-    end
+        puts "AirWatch Account Deactivation for User #{@invitation.recipient_username}, AirWatch ID #{airwatch_id}. Response Code: #{response.code}"
+      rescue => e
+        puts e
+      end
 
-    sleep 5 #Fix this. Currently have to do this because the AirWatch API can't handle deletion immediately after deactivation :(
+      sleep 5 #Fix this. Currently have to do this because the AirWatch API can't handle deletion immediately after deactivation :(
 
-    begin
-     	delete_url = "https://apple.awmdm.com/API/v1/system/users/#{airwatch_id}/delete"
-		response = RestClient::Request.execute( :method => :delete, 
-												:url => delete_url, 
-												:user => "#{api_user}", 
-												:password => "#{api_password}", 
-												:headers => { :host => "apple.awmdm.com", 
-															  :authorization => "Basic bW9oYW46bW9oYW4=", 
-															  'aw-tenant-code' => "#{aw_tenant_code}"})
-    	puts "AirWatch Account Deletion for User #{@invitation.recipient_username}, AirWatch ID #{airwatch_id}. Response Code: #{response.code}"    
-    rescue => e
-     	puts e
+      begin
+        delete_url = "https://apple.awmdm.com/API/v1/system/users/#{airwatch_id}/delete"
+      response = RestClient::Request.execute( :method => :delete, 
+                          :url => delete_url, 
+                          :user => "#{api_user}", 
+                          :password => "#{api_password}", 
+                          :headers => { :host => "apple.awmdm.com", 
+                                  :authorization => "Basic bW9oYW46bW9oYW4=", 
+                                  'aw-tenant-code' => "#{aw_tenant_code}"})
+        puts "AirWatch Account Deletion for User #{@invitation.recipient_username}, AirWatch ID #{airwatch_id}. Response Code: #{response.code}"    
+      rescue => e
+        puts e
+      end
     end
     #Done calling AirWatch API
+
+    if !airwatch_admin_user_id.nil?
+      begin
+        delete_url = "https://apple.awmdm.com/API/v1/system/admins/{airwatch_admin_id}/delete"
+      response = RestClient::Request.execute( :method => :delete, 
+                          :url => delete_url, 
+                          :user => "#{api_user}", 
+                          :password => "#{api_password}", 
+                          :headers => { :host => "apple.awmdm.com", 
+                                  :authorization => "Basic bW9oYW46bW9oYW4=", 
+                                  'aw-tenant-code' => "#{aw_tenant_code}"})
+        puts "AirWatch Admin Account Deletion for User #{@invitation.recipient_username}, AirWatch ID #{airwatch_id}. Response Code: #{response.code}"    
+      rescue => e
+        puts e
+      end
+    end
+    #Done calling AirWatch API
+
 
     puts "Calling ad replicate amer.."
     begin
